serving/requirements.txt
joblib==1.3.2
gunicorn==20.1.0
pillow==10.4.0
ultralytics==8.3.2
fastapi==0.115.0
uvicorn==0.31.1
python-multipart==0.0.12